requirements.txt
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.4.0
django-environ==0.4.5
django-mysql==3.8.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
matplotlib==3.3.2
mysqlclient==2.0.1
numpy==1.19.1
pydub==0.24.1
scipy==1.5.3